requirements.txt
numpy~=1.19.5
Pillow~=8.0.1
pytesseract~=0.3.6